requirements.txt
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9

# Dev dependencies
pytest>=8.0.0
pytest-mock>=3.14.0
responses>=0.25.0
